refactoring to isolate functionality and avoid circular dependencies
